requirements.txt
Flask==2.3.3
numpy==1.24.3
opencv-python==4.8.1.78
Pillow==10.0.1
redis==4.5.4
flask-caching==2.0.2
qrcode[pil]==7.4.2
Flask-Limiter==3.5.0
Flask-JWT-Extended==4.5.3
graphene==3.3
flask-graphql==2.0.1
graphql-core==3.2.3
cryptography==41.0.7
Flask-Talisman==1.0.0
flask-seasurf==0.3.1